refactor actions for the login API
- use createAction from redux-toolkit for brevity
- add comments about what the actions do
- use a common loginSagaFail action instead of a different one per saga,
  confusingly named to look like errors from the backend
- consolidate all the mostly one-line actions into a common
  loginActions.js
