--- src/login/redux/actions/postRefLoginActions.js
@@ -1,19 +1,0 @@
-export const POST_LOGIN_REF_TO_NEXT = "POST_LOGIN_REF_TO_NEXT";
-export const POST_IDP_NEXT_FAIL = "POST_IDP_NEXT_FAIL";
-
-export function useLoginRef() {
-  return {
-    type: POST_LOGIN_REF_TO_NEXT,
-  };
-}
-
-export function postRefFail(err) {
-  return {
-    type: POST_IDP_NEXT_FAIL,
-    error: true,
-    payload: {
-      message: err,
-    },
-  };
-}
-
