+++ src/login/redux/sagas/login/postRefLoginSaga.js
@@ -1,9 +1,10 @@
 import { call, select, put } from "redux-saga/effects";
 import postRequest from "../postDataRequest";
 import { putCsrfToken } from "../../../../sagas/common";
-import * as actions from "../../actions/postRefLoginActions";
+import * as actions from "../../actions/loginActions";
 import { eduidRMAllNotify } from "../../../../actions/Notifications";
 
+// Call the /next endpoint to see what the backend requires us to do to proceed with this login.
 export function* postRefLoginSaga() {
   const state = yield select((state) => state);
   const url = state.config.next_url;
@@ -19,6 +20,6 @@ export function* postRefLoginSaga() {
       yield put(eduidRMAllNotify());
     }
   } catch (error) {
-    yield put(actions.postRefFail(error.toString()));
+    yield put(actions.loginSagaFail(error.toString()));
   }
 }
