+++ src/login/redux/sagas/login/postTouVersionsSaga.js
@@ -1,14 +1,15 @@
 import { call, select, put } from "redux-saga/effects";
 import postRequest from "../postDataRequest";
 import { putCsrfToken } from "../../../../sagas/common";
-import * as actions from "../../actions/postTouVersionsActions";
+import { loginSagaFail } from "../../actions/loginActions";
 
+// Saga that posts the versions of the ToU that is available in this bundle to the /tou endpoint.
 export function* postTouVersionsSaga(action) {
   const state = yield select((state) => state);
   const url = state.login.post_to;
   const dataToSend = {
     ref: state.login.ref,
-    versions: action.payload.versions.toString(),
+    versions: action.payload.toString(),
     csrf_token: state.config.csrf_token,
   };
   try {
@@ -16,6 +17,6 @@ export function* postTouVersionsSaga(action) {
     yield put(putCsrfToken(response));
     yield put(response);
   } catch (error) {
-    yield put(actions.postTouVersionsFail(error.toString()));
+    yield put(loginSagaFail(error.toString()));
   }
 }
