+++ src/login/redux/sagas/login/postUpdatedTouAcceptSaga.js
@@ -1,9 +1,9 @@
 import { call, select, put } from "redux-saga/effects";
 import postRequest from "../postDataRequest";
 import { putCsrfToken } from "../../../../sagas/common";
-import * as actions from "../../actions/postUpdatedTouAcceptActions";
-import { useLoginRef } from "../../actions/postRefLoginActions";
+import { loginSagaFail, useLoginRef } from "../../actions/loginActions";
 
+// Saga to post the version of the ToU the user has just accepted to the /tou endpoint.
 export function* postUpdatedTouAcceptSaga(action) {
   const state = yield select((state) => state);
   const url = state.login.post_to;
@@ -20,6 +20,6 @@ export function* postUpdatedTouAcceptSaga(action) {
       yield put(useLoginRef());
     }
   } catch (error) {
-    yield put(actions.updateTouAcceptFail(error.toString()));
+    yield put(loginSagaFail(error.toString()));
   }
 }
