+++ src/tests/sagas/postRefForWebauthnChallengeSaga-test.js
@@ -2,7 +2,7 @@ import expect from "expect";
 import { call } from "redux-saga/effects";
 import postRequest from "../../login/redux/sagas/postDataRequest";
 import { postRefForWebauthnChallengeSaga } from "../../login/redux/sagas/login/postRefForWebauthnChallengeSaga";
-import { postRefForWebauthnChallengeFail } from "../../login/redux/actions/postRefForWebauthnChallengeActions";
+import { loginSagaFail } from "../../login/redux/actions/loginActions";
 
 const fakeState = {
   config: {
@@ -71,7 +71,7 @@ describe("first API call to /mfa behaves as expected on _FAIL", () => {
     expect(next.value.PUT.action.type).toEqual("NEW_CSRF_TOKEN");
     next = generator.next();
     expect(next.value.PUT.action.type).toEqual("POST_IDP_MFA_AUTH_FAIL");
-    expect(failResponse).toEqual(postRefForWebauthnChallengeFail("error"));
+    expect(failResponse).toEqual(loginSagaFail("error"));
   });
   it("done after 'POST_IDP_MFA_AUTH_FAIL'", () => {
     const done = generator.next().done;
