+++ src/tests/sagas/postTouVersionsSaga-test.js
@@ -2,7 +2,7 @@ import expect from "expect";
 import { call } from "redux-saga/effects";
 import postRequest from "../../login/redux/sagas/postDataRequest";
 import { postTouVersionsSaga } from "../../login/redux/sagas/login/postTouVersionsSaga";
-import { postTouVersionsFail } from "../../login/redux/actions/postTouVersionsActions";
+import { loginSagaFail } from "../../login/redux/actions/loginActions";
 
 const fakeState = {
   config: {
@@ -79,7 +79,7 @@ describe("first API call to /tou behaves as expected on _FAIL", () => {
     expect(next.value.PUT.action.type).toEqual("NEW_CSRF_TOKEN");
     next = generator.next();
     expect(next.value.PUT.action.type).toEqual("POST_IDP_TOU_FAIL");
-    expect(failResponse).toEqual(postTouVersionsFail("error"));
+    expect(failResponse).toEqual(loginSagaFail("error"));
   });
   it("done after 'POST_IDP_TOU_FAIL'", () => {
     const done = generator.next().done;
