+++ src/tests/sagas/postUpdatedTouAcceptSaga-test.js
@@ -2,7 +2,7 @@ import expect from "expect";
 import { call } from "redux-saga/effects";
 import postRequest from "../../login/redux/sagas/postDataRequest";
 import { postUpdatedTouAcceptSaga } from "../../login/redux/sagas/login/postUpdatedTouAcceptSaga";
-import { updateTouAcceptFail } from "../../login/redux/actions/postUpdatedTouAcceptActions";
+import { loginSagaFail } from "../../login/redux/actions/loginActions";
 
 const fakeState = {
   config: {
@@ -84,7 +84,7 @@ describe("second API call to /tou behaves as expected on _FAIL", () => {
     expect(next.value.PUT.action.type).toEqual("NEW_CSRF_TOKEN");
     next = generator.next();
     expect(next.value.PUT.action.type).toEqual("POST_IDP_TOU_FAIL");
-    expect(failResponse).toEqual(updateTouAcceptFail("error"));
+    expect(failResponse).toEqual(loginSagaFail("error"));
   });
   it("done after 'POST_IDP_TOU_FAIL'", () => {
     const done = generator.next().done;
